requirements.txt
fastapi
uvicorn
streamlit
requests
pandas
python-dotenv
fastmcp
langchain-google-genai
langchain-core
pydantic
langsmith
